fix, compare equally-signed values

+++ src/utetpc.c
@@ -970,7 +970,7 @@ seek_sort(uteseek_t sk)
 	size_t sk_sz = seek_size(sk);
 
 	/* we never hand out bigger pages */
-	assert(sk_sz <= sysconf(_SC_PAGESIZE) * 1024U);
+	assert(sk_sz <= (size_t)sysconf(_SC_PAGESIZE) * 1024U);
 
 	/* get us another map */
 	new = mmap(NULL, sizeof(*new) + sk_sz, PROT_MEM, MAP_MEM, -1, 0);
